move prefetch serialization tests to kernel tests
