+++ tensorflow/python/data/kernel_tests/prefetch_test.py
@@ -21,6 +21,7 @@ import time
 
 from absl.testing import parameterized
 
+from tensorflow.python.data.kernel_tests import checkpoint_test_base
 from tensorflow.python.data.kernel_tests import test_base
 from tensorflow.python.data.ops import dataset_ops
 from tensorflow.python.framework import combinations
@@ -74,6 +75,17 @@ class PrefetchTest(test_base.DatasetTestBase, parameterized.TestCase):
       sess.close()
       thread.join()
 
+class PrefetchDatasetCheckpointTest(checkpoint_test_base.CheckpointTestBase,
+                                    parameterized.TestCase):
+
+  def build_dataset(self, seed):
+    return dataset_ops.Dataset.range(100).prefetch(10).shuffle(
+        buffer_size=10, seed=seed, reshuffle_each_iteration=False)
+
+  @combinations.generate(test_base.default_test_combinations())
+  def testCore(self):
+    num_outputs = 100
+    self.run_core_tests(lambda: self.build_dataset(10), num_outputs)
 
 if __name__ == "__main__":
   test.main()
